Furniture Gallery
A web-based project called "Furniture Gallery," developed as an assignment from Mr. Nabi Bux. This project showcases a range of furniture items, built with a focus on both functionality and visual appeal. Completed in one week with passion and dedication, this project represents my commitment to creating an engaging and user-friendly interface for browsing furniture collections. It highlights key features like easy navigation, detailed item descriptions, and a responsive design.